src/main/java/com/lendico/plan/web/serializer/CustomToStringSerializer.java
package com.lendico.plan.web.serializer;

import com.fasterxml.jackson.core.JsonGenerator;
import com.fasterxml.jackson.databind.JsonSerializer;
import com.fasterxml.jackson.databind.SerializerProvider;
import lombok.extern.slf4j.Slf4j;

import java.io.IOException;

@Slf4j
public class CustomToStringSerializer extends JsonSerializer<Object> {

    @Override
    public void serialize(Object value, JsonGenerator gen, SerializerProvider arg2) throws IOException {
        if (value == null) {
            // do nothing
        } else {
            gen.writeObject(value.toString());
        }
    }
}
